feat: use -Data instead of .@ for content
Received some feedback (from Kurt) that the .@ convention is
somewhat jarring, and not easy to grasp right away.

So, to reduce the friction, I am changing the container to be
defined by a field that starts with a dash.

It could be anything you want, e.g. -Data, -Value, -Content, -Etc.

Now, for the actual xml2yaml tool, I chose -Data as the name "Data"
could imply either pure char-data, or actual children elements,
so it works on both scenarios.