fix: modify install script to prioritize stable releases
Before, the script was picking whatever was the most recent released tag.
This is not necessarily correct, as you can have tags for betas, and release candidates.
We don't want customers installing the "latest" and end up using an unstable build.

Now, we sort all the tags by the semantic version number, and then
we pick the first stable one (not containing a dash).
If we can't find any stable one, we pick the top one in the list.