feat: add support for server-side tool filter on MCP API proxy
This feature allows client to pass a header `x-mcp-tools-filter`
when calling the `tools/list`. The filter is a comma separated
list of tools that the client wants to see in the tools/list response.

This is useful for showing only a subset of tools to an LLM and avoiding
context rot.

Also, modify the implementation of for API product authorization based filtering to
use a simple comma separated list as well, and be consistent with tools filtering.