bug: use generic []any and map[string]any in --set flags
The --set flags where using Map and Slice custom structs to set values in the context.

Because of this, the dig function did not always work well when trying to grab elements
from deep in the $.Values context.

Now, the internal structures used by the --set flags are always []any or map[string]any

It should now be possible to use the sprig dig function to get elements that are deep.