feat: add support for Terraform HCL format
With the new transform commands you should be able to go back and forth
between Terraform HCL and JSON. The transform between these two formats
adheres to the rules specified at:
  https://developer.hashicorp.com/terraform/language/syntax/json

  * transform tf-to-json
    Converts input Terraform HCL into HCL JSON
  * transform json-to-tf
    Converts input JSON into Terraform HCL

There is also a new flag (--set-tf) that can be passed to the render commands.
This flag works the same way as the other flags in the same family.

  --set-tf context_key=./path/to/terraform/file.tf

The input Terraform HCL file is converted to JSON format and made
available in the context as a map[string]any.